allen_academy_register/requirements.txt
Django==4.2.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
python-dotenv==1.0.1
python_dateutil==2.9.0.post0
bcrypt==4.1.2
